default lang on get value

+++ src/Jrmadsen67/MahanaTranslationManager/repositories/TranslationManagerEloquentRepository.php
@@ -16,7 +16,7 @@ class TranslationManagerEloquentRepository implements TranslationManagerReposito
 	}
 	public function update($key, $lang_code, $data, $cascade)
 	{
-		$translation = self::getValue($key, $lang_code);
+		$translation = self::find($key, $lang_code);
 		$translation->fill($data);
 		$translation->save();
 
@@ -31,31 +31,42 @@ class TranslationManagerEloquentRepository implements TranslationManagerReposito
 	public function delete($key, $lang_code)
 	{
 		$translation = self::getValue($key, $lang_code);
+		if (empty($translation)) {return false;}
 		$translation->delete();
-
 		return true;
 	}
 
 
 	public function getValue($key, $lang_code)
 	{
-		return self::find($key, $lang_code);
+		$value = self::find($key, $lang_code);
+
+		if (empty($value))
+		{
+			$value = self::find($key, \Config::get('mahana-translation-manager::translation_manager.default_lang'));
+		}	
+		return $value;
 	}
 	public function getLanguageSet($lang_code)
 	{
 		return Translations::whereLangCode($lang_code)->get();
 	}
+
 	public function getNeedTranslation($lang_code)
 	{
-
+		return Translations::whereLangCode(\Config::get('mahana-translation-manager::translation_manager.default_lang'))
+			->whereRaw('`key` NOT IN (SELECT t2.key FROM '.
+				\Config::get('mahana-translation-manager::translation_manager.table_name')
+				.' t2 WHERE t2.lang_code = "'.$lang_code.'")')
+		    ->get();
 	}
 	public function getNeedUpdate($lang_code)
 	{
-		return Translations::whereLangCode($lang_code)->whereRequireUpdate(1)->get();
+		return Translations::whereLangCode($lang_code)->whereRequiresUpdate(1)->get();
 	}
 	public function getNeedManualtranslation($lang_code)
 	{
-		return Translations::whereLangCode($lang_code)->whereRequireManualTranslation(1)->get();
+		return Translations::whereLangCode($lang_code)->whereRequiresManualTranslation(1)->get();
 	}
